Extract 'bookingActions' module for handling links based on role
We start by showing that the `WORKFLOW_MANAGER` (a role used by
the x-service CRU team) is given the action to "Withdraw". We'll
be moving to a new role (suggested `CRU_TEAM`) for actions which live
within the "Manage" section but aren't for AP Staff (`*MANAGER` roles).

See this table for the roles which should see each of the booking actions:

| Action             | Roles                   | notes                                                   |
| ------------------ | ----------------------- | ------------------------------------------------------- |
| move person        | LEGACY_MANAGER, MANAGER | NE users of "v1 manage"                                 |
| change dates       | LEGACY_MANAGER, MANAGER | NE users of "v1 manage"                                 |
| arrived            | MANAGER                 | deprecated feature, will return improved in "v2 manage" |
| not arrived        | MANAGER                 | deprecated feature, will return improved in "v2 manage" |
| departed           | MANAGER                 | deprecated feature, will return improved in "v2 manage" |
| withdraw placement | WORKFLOW_MANAGER        | we'll switch to CRU_TEAM                                |
