+++ server/utils/bookings/index.ts
@@ -1,12 +1,10 @@
 import type {
   BespokeError,
-  IdentityBarMenu,
   SelectOption,
   SummaryListItem,
   SummaryListWithCard,
   TableCell,
   TableRow,
-  UserDetails,
 } from '@approved-premises/ui'
 import type { BedSummary, Booking, BookingStatus, BookingSummary, PremisesBooking } from '@approved-premises/api'
 import { addDays, isBefore, isEqual, isWithinInterval } from 'date-fns'
@@ -168,94 +166,6 @@ export const bookingsToTableRows = (
 
 export const nameCell = (booking: PremisesBooking): TableCell =>
   isFullPerson(booking.person) ? { text: laoName(booking.person) } : { text: `LAO: ${booking.person.crn}` }
-
-export const bookingActions = (user: UserDetails, booking: Booking): Array<IdentityBarMenu> => {
-  if (user.roles?.includes('future_manager')) return v2BookingActions(booking)
-  if (user.roles?.includes('workflow_manager')) return legacyBookingActions(booking)
-  return []
-}
-
-export const v2BookingActions = (booking: Booking): Array<IdentityBarMenu> => {
-  if (booking.status === 'awaiting-arrival' && booking?.applicationId)
-    return [
-      {
-        items: [
-          {
-            text: 'Withdraw placement',
-            classes: 'govuk-button--secondary',
-            href: applyPaths.applications.withdraw.new({ id: booking.applicationId }),
-          },
-        ],
-      },
-    ]
-
-  return []
-}
-
-export const legacyBookingActions = (booking: Booking): Array<IdentityBarMenu> => {
-  const withdrawalLink = !booking?.applicationId
-    ? paths.bookings.cancellations.new({ premisesId: booking.premises.id, bookingId: booking.id })
-    : applyPaths.applications.withdraw.new({ id: booking?.applicationId })
-
-  if (booking.status === 'awaiting-arrival' || booking.status === 'arrived') {
-    const items = [
-      {
-        text: 'Move person to a new bed',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.moves.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      },
-    ]
-
-    if (booking.status === 'awaiting-arrival') {
-      items.push({
-        text: 'Mark as arrived',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.arrivals.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      })
-      items.push({
-        text: 'Mark as not arrived',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.nonArrivals.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      })
-      items.push({
-        text: 'Withdraw placement',
-        classes: 'govuk-button--secondary',
-        href: withdrawalLink,
-      })
-      items.push({
-        text: 'Change placement dates',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.dateChanges.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      })
-    }
-
-    if (booking.status === 'arrived') {
-      items.push({
-        text: 'Log departure',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.departures.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      })
-      items.push({
-        text: 'Update departure date',
-        classes: 'govuk-button--secondary',
-        href: paths.bookings.extensions.new({ premisesId: booking.premises.id, bookingId: booking.id }),
-      })
-      items.push({
-        text: 'Withdraw placement',
-        classes: 'govuk-button--secondary',
-        href: withdrawalLink,
-      })
-    }
-
-    return [
-      {
-        items,
-      },
-    ]
-  }
-
-  return null
-}
 
 export const generateConflictBespokeError = (
   err: SanitisedError,
